+++ x/auth/tx/service.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client/grpc/tmservice"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/cosmos/cosmos-sdk/types/query"
 
 	gogogrpc "github.com/gogo/protobuf/grpc"
 	"github.com/golang/protobuf/proto" // nolint: staticcheck
@@ -52,9 +53,16 @@ func (s txServer) GetTxsEvent(ctx context.Context, req *txtypes.GetTxsEventReque
 		return nil, status.Error(codes.InvalidArgument, "request cannot be nil")
 	}
 
-	page, limit, err := pagination.ParsePagination(req.Pagination)
-	if err != nil {
-		return nil, err
+	page := int(req.Page)
+	// Tendermint node.TxSearch that is used for querying txs defines pages starting from 1,
+	// so we default to 1 if not provided in the request.
+	if page == 0 {
+		page = 1
+	}
+
+	limit := int(req.Limit)
+	if limit == 0 {
+		limit = query.DefaultLimit
 	}
 	orderBy := parseOrderBy(req.OrderBy)
 
@@ -88,9 +96,7 @@ func (s txServer) GetTxsEvent(ctx context.Context, req *txtypes.GetTxsEventReque
 	return &txtypes.GetTxsEventResponse{
 		Txs:         txsList,
 		TxResponses: result.Txs,
-		Pagination: &pagination.PageResponse{
-			Total: result.TotalCount,
-		},
+		Total:       result.TotalCount,
 	}, nil
 }